Bug fixes including #210

+++ lazylibrarian/webServe.py
@@ -375,11 +375,12 @@ class WebInterface(object):
     def deleteAuthor(self, AuthorID):
         myDB = database.DBConnection()
         authorsearch = myDB.select('SELECT AuthorName from authors WHERE AuthorID="%s"' % AuthorID)
-        AuthorName = authorsearch[0]['AuthorName']
-        logger.info(u"Removing all references to author: %s" % AuthorName)
+        if len(authorsearch):  # to stop error if try to delete an author while they are still loading
+            AuthorName = authorsearch[0]['AuthorName']
+            logger.info(u"Removing all references to author: %s" % AuthorName)
 
-        myDB.action('DELETE from authors WHERE AuthorID="%s"' % AuthorID)
-        myDB.action('DELETE from books WHERE AuthorID="%s"' % AuthorID)
+            myDB.action('DELETE from authors WHERE AuthorID="%s"' % AuthorID)
+            myDB.action('DELETE from books WHERE AuthorID="%s"' % AuthorID)
         raise cherrypy.HTTPRedirect("home")
     deleteAuthor.exposed = True
 
@@ -463,13 +464,15 @@ class WebInterface(object):
             find_book.join()
 
         books = []
-        mags = False
         books.append({"bookid": bookid})
 
+        
         if (lazylibrarian.USE_NZB):
-            threading.Thread(target=search_nzb_book, args=[books, mags]).start()
+            threading.Thread(target=search_nzb_book, args=[books]).start()
         if (lazylibrarian.USE_TOR):
-            threading.Thread(target=search_tor_book, args=[books, mags]).start()
+            threading.Thread(target=search_tor_book, args=[books]).start()
+        if not lazylibrarian.USE_NZB and not lazylibrarian.USE_TOR:
+            logger.warn("No search methods set, check config.")
 
         raise cherrypy.HTTPRedirect("books")
     addBook.exposed = True
@@ -500,13 +503,14 @@ class WebInterface(object):
             books = []
             books.append({"bookid": bookid})
 
-            mags = False
             if (lazylibrarian.USE_NZB):
-                threading.Thread(target=search_nzb_book, args=[books, mags]).start()
+                threading.Thread(target=search_nzb_book, args=[books]).start()
             if (lazylibrarian.USE_TOR):
-                threading.Thread(target=search_tor_book, args=[books, mags]).start()
-
-            logger.debug(u"Searching for book with id: " + bookid)
+                threading.Thread(target=search_tor_book, args=[books]).start()
+            if not lazylibrarian.USE_NZB and not lazylibrarian.USE_TOR:
+                logger.warn(u"No search methods set, check config.")
+            else:
+                logger.debug(u"Searching for book with id: " + bookid)
         if AuthorName:
             raise cherrypy.HTTPRedirect("authorPage?AuthorName=%s" % AuthorName)
     searchForBook.exposed = True
@@ -570,11 +574,13 @@ class WebInterface(object):
                 # ouch dirty workaround...
                 if not bookid == 'book_table_length':
                     books.append({"bookid": bookid})
-            mags = False
+
             if (lazylibrarian.USE_NZB):
-                threading.Thread(target=search_nzb_book, args=[books, mags]).start()
+                threading.Thread(target=search_nzb_book, args=[books]).start()
             if (lazylibrarian.USE_TOR):
-                threading.Thread(target=search_tor_book, args=[books, mags]).start()
+                threading.Thread(target=search_tor_book, args=[books]).start()
+            if not lazylibrarian.USE_NZB and not lazylibrarian.USE_TOR:
+                logger.warn(u"No search methods set, check config.")
 
         if redirect == "author":
             raise cherrypy.HTTPRedirect("authorPage?AuthorName=%s" % AuthorName)
@@ -882,22 +888,23 @@ class WebInterface(object):
     def showJobs(self):
         # show the current status of LL cron jobs in the log
         for job in lazylibrarian.SCHED.get_jobs():
-            if "search_magazines" in str(job):
+            job = str(job)
+            if "search_magazines" in job:
                 jobname = "[CRON] - Check for new magazine issues"
-            elif "checkForUpdates" in str(job):
+            elif "checkForUpdates" in job:
                 jobname = "[CRON] - Check for LazyLibrarian update"
-            elif "search_tor_book" in str(job):
+            elif "search_tor_book" in job:
                 jobname = "[CRON] - TOR book search"
-            elif "search_nzb_book" in str(job):
+            elif "search_nzb_book" in job:
                 jobname = "[CRON] - NZB book search"
-            elif "processDir" in str(job):
+            elif "processDir" in job:
                 jobname = "[CRON] - Process download directory"
             else:       
-                jobname = str(job).split(' ')[0].split('.')[2]
+                jobname = job.split(' ')[0].split('.')[2]
 
-            jobtime = str(job).split('[')[1].split('.')[0]
+            jobtime = job.split('[')[1].split('.')[0]
             logger.info(u"%s [%s" % (jobname, jobtime))
-        logger.info(u"XMLCache %s hits, %s miss" % (int(lazylibrarian.CACHE_HIT), int(lazylibrarian.CACHE_MISS)))
+        logger.info(u"XMLCache %i hits, %i miss" % (int(lazylibrarian.CACHE_HIT), int(lazylibrarian.CACHE_MISS)))
         raise cherrypy.HTTPRedirect("logs")
     showJobs.exposed = True
 
@@ -995,14 +1002,14 @@ class WebInterface(object):
             return serve_template(templatename="history.html", title="History", history=history)
     history.exposed = True
 
-    def clearhistory(self, type=None):
+    def clearhistory(self, histtype=None):
         myDB = database.DBConnection()
-        if type == 'all':
+        if histtype == 'all':
             logger.info(u"Clearing all history")
             myDB.action("DELETE from wanted WHERE Status != 'Skipped' and Status != 'Ignored'")
         else:
-            logger.info(u"Clearing history where status is %s" % type)
-            myDB.action('DELETE from wanted WHERE Status="%s"' % type)
+            logger.info(u"Clearing history where status is %s" % histtype)
+            myDB.action('DELETE from wanted WHERE Status="%s"' % histtype)
         raise cherrypy.HTTPRedirect("history")
     clearhistory.exposed = True
 
@@ -1090,6 +1097,8 @@ class WebInterface(object):
                 threading.Thread(target=search_nzb_book).start()
             if (lazylibrarian.USE_TOR):
                 threading.Thread(target=search_tor_book).start()
+            if not lazylibrarian.USE_NZB and not lazylibrarian.USE_TOR:
+                logger.warn(u"No search methods set, check config.")
         raise cherrypy.HTTPRedirect(source)
     forceSearch.exposed = True
